+++ pysat/constellations/testing_empty.py
@@ -1,5 +1,11 @@
 """
 Creates an empty constellation for testing
+
+Attributes
+----------
+instruments : list
+    List of pysat.Instrument objects
+
 """
 
 instruments = []